refactor(database): extract the connection-priority layer out of index.ts
The DB-connection-priority feature had grown ~10 functions inside
database/index.ts (default/extra connection names, base-config
namespacing, priority, uniqueness assertion, the named-connection
registry, and request resolution). An index.ts is an entrypoint, not a
home for feature logic, so move all of it to a sibling connections.ts.

index.ts keeps the default-pool singleton, knex construction, and the
validators; it imports getBaseDbConfig/assertConnectionNamesAreUnique
from connections and re-exports the two public *ForAccountability
functions so no caller changes. Pure structure move, no behaviour change.
